pyproject.toml
[tool.poetry]
name = "gcode_sender"
version = "0.1.0"
description = ""
authors = ["Jonas Nogueira <jonasjrsantos@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
